+++ packages/tools/src/store/state.ts
@@ -55,3 +55,5 @@ export {
   state,
   state as default,
 };
+
+window.state = state;
