+++ packages/tools/src/tools/base/AnnotationTool.ts
@@ -9,7 +9,7 @@ import type { Types } from '@cornerstonejs/core';
 
 import { vec4, vec2 } from 'gl-matrix';
 
-import BaseTool from './BaseTool';
+import AnnotationDisplayTool from './AnnotationDisplayTool';
 import { isAnnotationLocked } from '../../stateManagement/annotation/annotationLocking';
 import { isAnnotationVisible } from '../../stateManagement/annotation/annotationVisibility';
 import { getViewportSpecificAnnotationManager } from '../../stateManagement/annotation/annotationState';
@@ -37,7 +37,7 @@ import { StyleSpecifier } from '../../types/AnnotationStyle';
  * To create a new annotation tool, derive from this class and implement the
  * abstract methods.
  */
-abstract class AnnotationTool extends BaseTool {
+abstract class AnnotationTool extends AnnotationDisplayTool {
   static toolName;
   // ===================================================================
   // Abstract Methods - Must be implemented.
@@ -53,18 +53,6 @@ abstract class AnnotationTool extends BaseTool {
     evt: EventTypes.MouseDownActivateEventType,
     interactionType: InteractionTypes
   ): Annotation;
-
-  /**
-   * @abstract renderAnnotation it used to draw the tool's annotation in each
-   * request animation frame
-   *
-   * @param enabledElement - The Cornerstone's enabledElement.
-   * @param svgDrawingHelper - The svgDrawingHelper providing the context for drawing.
-   */
-  abstract renderAnnotation(
-    enabledElement: Types.IEnabledElement,
-    svgDrawingHelper: SVGDrawingHelper
-  );
 
   /**
    * @abstract cancel Used to cancel the ongoing tool drawing and manipulation
@@ -118,31 +106,6 @@ abstract class AnnotationTool extends BaseTool {
     proximity: number,
     interactionType: string
   ): boolean;
-
-  /**
-   * @virtual Given the element and annotations which is an array of annotation, it
-   * filters the annotations array to only include the annotation based on the viewportType.
-   * If the viewport is StackViewport, it filters based on the current imageId of the viewport,
-   * if the viewport is volumeViewport, it only returns those that are within the
-   * same slice as the current rendered slice in the volume viewport.
-   * imageId as the enabledElement.
-   * @param element - The HTML element
-   * @param annotations - The annotations to filter (array of annotation)
-   * @returns The filtered annotations
-   */
-  filterInteractableAnnotationsForElement(
-    element: HTMLDivElement,
-    annotations: Annotations
-  ): Annotations | undefined {
-    if (!annotations || !annotations.length) {
-      return;
-    }
-
-    const enabledElement = getEnabledElement(element);
-    const { viewport } = enabledElement;
-
-    return filterAnnotationsForDisplay(viewport, annotations);
-  }
 
   /**
    * @virtual Event handler for Cornerstone MOUSE_MOVE event.
@@ -205,105 +168,6 @@ abstract class AnnotationTool extends BaseTool {
   };
 
   /**
-   * On Image Calibration, take all the annotation from the AnnotationState manager,
-   * and invalidate them to force them to be re-rendered and their stats to be recalculated.
-   * Then use the old and new imageData (non-calibrated and calibrated) to calculate the
-   * new position for the annotations in the space of the new imageData.
-   *
-   * @param evt - The calibration event
-   *
-   */
-  public onImageSpacingCalibrated = (
-    evt: Types.EventTypes.ImageSpacingCalibratedEvent
-  ) => {
-    const {
-      element,
-      rowScale,
-      columnScale,
-      imageId,
-      imageData: calibratedImageData,
-      worldToIndex: noneCalibratedWorldToIndex,
-    } = evt.detail;
-
-    const { viewport } = getEnabledElement(element);
-
-    if (viewport instanceof VolumeViewport) {
-      throw new Error('Cannot calibrate a volume viewport');
-    }
-
-    const calibratedIndexToWorld = calibratedImageData.getIndexToWorld();
-
-    const imageURI = utilities.imageIdToURI(imageId);
-    const stateManager = getViewportSpecificAnnotationManager(element);
-    const framesOfReference = stateManager.getFramesOfReference();
-
-    // For each frame Of Reference
-    framesOfReference.forEach((frameOfReference) => {
-      const frameOfReferenceSpecificAnnotations =
-        stateManager.getFrameOfReferenceAnnotations(frameOfReference);
-
-      const toolSpecificAnnotations =
-        frameOfReferenceSpecificAnnotations[this.getToolName()];
-
-      if (!toolSpecificAnnotations || !toolSpecificAnnotations.length) {
-        return;
-      }
-
-      // for this specific tool
-      toolSpecificAnnotations.forEach((annotation) => {
-        // if the annotation is drawn on the same imageId
-        const referencedImageURI = utilities.imageIdToURI(
-          annotation.metadata.referencedImageId
-        );
-
-        if (referencedImageURI === imageURI) {
-          // make them invalid since the image has been calibrated so that
-          // we can update the cachedStats and also rendering
-          annotation.invalidated = true;
-          annotation.data.cachedStats = {};
-
-          // Update annotation points to the new calibrated points. Basically,
-          // using the worldToIndex function we get the index on the non-calibrated
-          // image and then using the calibratedIndexToWorld function we get the
-          // corresponding point on the calibrated image world.
-          annotation.data.handles.points = annotation.data.handles.points.map(
-            (point) => {
-              const p = vec4.fromValues(...point, 1);
-              const pCalibrated = vec4.fromValues(0, 0, 0, 1);
-              const nonCalibratedIndexVec4 = vec4.create();
-              vec4.transformMat4(
-                nonCalibratedIndexVec4,
-                p,
-                noneCalibratedWorldToIndex
-              );
-              const calibratedIndex = [
-                columnScale * nonCalibratedIndexVec4[0],
-                rowScale * nonCalibratedIndexVec4[1],
-                nonCalibratedIndexVec4[2],
-              ];
-
-              vec4.transformMat4(
-                pCalibrated,
-                vec4.fromValues(
-                  calibratedIndex[0],
-                  calibratedIndex[1],
-                  calibratedIndex[2],
-                  1
-                ),
-                calibratedIndexToWorld
-              );
-
-              return pCalibrated.slice(0, 3) as Types.Point3;
-            }
-          );
-        }
-      });
-
-      triggerAnnotationRender(element);
-    });
-  };
-
-  /**
    * It checks if the mouse click is near TextBoxHandle or AnnotationHandle itself, and
    * return either it. It prioritize TextBoxHandle over AnnotationHandle. If
    * the mouse click is not near any of the handles, it does not return anything.
@@ -362,59 +226,6 @@ abstract class AnnotationTool extends BaseTool {
     }
 
     data.handles.activeHandleIndex = null;
-  }
-
-  protected getReferencedImageId(
-    viewport: Types.IStackViewport | Types.IVolumeViewport,
-    worldPos: Types.Point3,
-    viewPlaneNormal: Types.Point3,
-    viewUp: Types.Point3
-  ): string {
-    const targetId = this.getTargetId(viewport);
-
-    let referencedImageId;
-
-    if (viewport instanceof StackViewport) {
-      referencedImageId = targetId.split('imageId:')[1];
-    } else {
-      const volumeId = targetId.split('volumeId:')[1];
-      const imageVolume = cache.getVolume(volumeId);
-
-      referencedImageId = utilities.getClosestImageId(
-        imageVolume,
-        worldPos,
-        viewPlaneNormal,
-        viewUp
-      );
-    }
-
-    return referencedImageId;
-  }
-
-  /**
-   * It takes the property (color, lineDash, etc.) and based on the state of the
-   * annotation (selected, highlighted etc.) it returns the appropriate value
-   * based on the central toolStyle settings for each level of specification.
-   * @param property - The name of the style property to get.
-   * @param styleSpecifier - An object containing the specifications such as viewportId,
-   * toolGroupId, toolName and annotationUID which are used to get the style if the level of specificity is
-   * met (hierarchy is checked from most specific to least specific which is
-   * annotationLevel -> viewportLevel -> toolGroupLevel -> default.
-   * @param annotation - The annotation for the tool that is
-   * currently active.
-   * @returns The value of the property.
-   */
-  public getStyle(
-    property: string,
-    specifications: StyleSpecifier,
-    annotation?: Annotation
-  ): unknown {
-    return getStyleProperty(
-      property,
-      specifications,
-      getState(annotation),
-      this.mode
-    );
   }
 
   /**
